track video sources rather than scenes
allows tally light to respond correctly when a camera is used in multiple
scenes. example: if the watched source is in the program scene, display the
live/program color.

also monitor for events which update scene composition:

  * adding/removing sources.
  * making sources visible/invisible.
  * these changes will affect the tally light only if the watched source is
    not currently live. (when making changes to a live scene, OBS will only
    apply those changes in preview. changing the current program scene requires
    a transition. so when we are live, we don't update the tally light except
    when the scene is actually changed.)

i think this code could probably be simplified somewhat. (feels a little messy.)

the most surprising cases i hit were when:

  * 2 cameras are in use.
  * 2 tally lights are active. 1 tracking each camera.
  * a single scene contains both cameras.
  * edits are made to the scene - making cameras visible/invisible, or
    adding/removing a camera from the scene.

that prompted the introduction of the `canUpdateIfInProgram` variable. it works,
but i suspect modeling the data differently might allow these kinds of tracking
variables to be simplified at some point.